refactor: move package to authentication
